Retain fractional milliseconds in RTT measurements
Change the clock callback to return double milliseconds through an output pointer, making the API change explicit. Retain precise send and receive timestamps alongside the integer clock used for scheduling, and use float RTT samples and minima throughout congestion control and pacing.

Keep the 1ms minimum RTT, ACK-delay exponent 10, and integer PTO durations. Preserve fractional ACK-delay adjustment and round absolute loss, closing, and draining deadlines up when scheduling millisecond timers. Use infinity for timestamp sentinels and NaN for the unlocked cached clock.

Update clock implementations, callers, and tracing. Add regressions for fractional RTT and ACK delay, timer boundaries, pacing, rapid start, and careful resume.

Validation: test.t, CLI, simulator, and echo example build; all 33 unit-test groups pass with lossy expectations unchanged. CLI e2e has failures in path migration, slow-start timing, reset-stream-overflow, and invalid-ACK groups (4 of 26); those expectations are unchanged.

+++ lib/cc-pico.c
@@ -335,7 +335,7 @@ static uint32_t cubic_update_w_est(struct st_quicly_cc_cubic_t *state, uint32_t 
 }
 
 static void cubic_on_acked(struct st_quicly_cc_cubic_t *state, uint32_t *cwnd, uint32_t cwnd_epoch, uint32_t bytes, int cc_limited,
-                           uint32_t rtt, int64_t now, uint32_t actual_mtu, uint32_t reference_mtu)
+                           float rtt, int64_t now, uint32_t actual_mtu, uint32_t reference_mtu)
 {
     uint32_t w_est = cubic_quantized_w_est(state, cwnd_epoch, actual_mtu);
     if (cc_limited)
